move socket code to separate place
